add doc to tests
